add: methonds Users and Delete

+++ pkg/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	Id       int    `json:"-" db:"id"`
+	Id       uint64 `json:"-" db:"id"`
 	Name     string `json:"name" db:"name"`
 	Email    string `json:"email" db:"email"`
 	Password string `json:"password" db:"password"`
