+++ pkg/repository/repository.go
@@ -12,7 +12,7 @@ type Repository struct {
 type User interface {
 	Create(user *entity.User) (uint64, error)
 	GetAll() ([]entity.User, error)
-	Delete(userId int) error
+	Delete(userId uint64) error
 }
 
 func NewRepository(db *sql.DB) *Repository {
